+++ packages/edition-slices/edition-slices.showcase.js
@@ -46,13 +46,29 @@ const publications = {
   ST: "SUNDAYTIMES",
   TIMES: "TIMES"
 };
+
+const onArticleSavePress = {
+  onArticleSavePressNotPresent: null,
+  onArticleSavePressPresent: () => {}
+};
+const savedArticles = {
+  savedArticlesNotPresent: null,
+  savedArticlesPresent: { 1: true, 2: true }
+};
+
 /* eslint-disable react/prop-types */
 const renderSlice = (Component, data) => ({ select }, { decorateAction }) => (
   <Responsive>
     <ScrollView>
       <SectionContext.Provider
         value={{
-          publicationName: select("Publication:", publications, "TIMES")
+          onArticleSavePress: select(
+            "onArticleSavePress:",
+            onArticleSavePress,
+            null
+          ),
+          publicationName: select("Publication:", publications, "TIMES"),
+          savedArticles: select("savedArticles:", savedArticles, null)
         }}
       >
         <Component