+++ packages/edition-slices/src/tiles/tile-i/index.js
@@ -18,6 +18,7 @@ const TileI = ({ onPress, tile }) => (
     />
     <TileSummary
       headlineStyle={styles.headline}
+      starStyle={styles.star}
       style={styles.summaryContainer}
       tile={tile}
     />