+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.15](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@1.1.14...@times-components/article-magazine-standard@1.1.15) (2019-04-11)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [1.1.14](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@1.1.13...@times-components/article-magazine-standard@1.1.14) (2019-04-11)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard