+++ packages/caption/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.13](https://github.com/newsuk/times-components/compare/@times-components/caption@3.0.12...@times-components/caption@3.0.13) (2019-04-11)
+
+**Note:** Version bump only for package @times-components/caption
+
+
+
+
+
 ## [3.0.12](https://github.com/newsuk/times-components/compare/@times-components/caption@3.0.11...@times-components/caption@3.0.12) (2019-04-11)
 
 **Note:** Version bump only for package @times-components/caption