+++ packages/star-button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.1.12...@times-components/star-button@0.1.13) (2019-04-11)
+
+**Note:** Version bump only for package @times-components/star-button
+
+
+
+
+
 ## [0.1.12](https://github.com/newsuk/times-components/compare/@times-components/star-button@0.1.11...@times-components/star-button@0.1.12) (2019-04-11)
 
 **Note:** Version bump only for package @times-components/star-button