+++ packages/test-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.27](https://github.com/newsuk/times-components/compare/@times-components/test-utils@2.2.26...@times-components/test-utils@2.2.27) (2019-04-11)
+
+**Note:** Version bump only for package @times-components/test-utils
+
+
+
+
+
 ## [2.2.26](https://github.com/newsuk/times-components/compare/@times-components/test-utils@2.2.25...@times-components/test-utils@2.2.26) (2019-04-11)
 
 **Note:** Version bump only for package @times-components/test-utils